Tree Pruning
Given a tree with NN vertices numbered from 11 to NN. The vertex 11 is the root of the tree. Each vertex is assigned with an integer weight. A remove operation can remove sub-tree rooted at an arbitrary vertex uu. You must use at most KK remove operations (possibly zero) so that the total weight of the remaining vertices is largest.

Input Format 
The first line contains two integers N,KN,K. 
The second line contains NN integers, the ithith integer is the weight of the ithith vertex. 
Each of the next N−1N−1 lines contains a pair of integers uu and vv, which represents an edge from vertex uu to vertex vv.

Output Format 
Print out a single integer which is the largest total weight of the remaining vertices.

Constraints 
2≤N≤1052≤N≤105 
1≤K≤2001≤K≤200 
−109≤weight≤109